pyproject.toml
[tool.poetry]
name = "astrosoft"
version = "0.1.0"
description = ""
authors = ["Philippe Gagnon <12717218+pgagnon@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.9"
click = "*"
cookiecutter = "*"
apache-airflow = "*"
apache-airflow-providers-postgres = "*"
apache-airflow-providers-mysql = "*"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
poetry = 'astrosoft.main:run'
